chore: refactor reflection
