+++ src/modules/text/Panel.tsx
@@ -1,7 +1,7 @@
-import { ActionIcon, Box, Checkbox, Select, Stack } from '@mantine/core';
+import { ActionIcon, Box, Select, Stack } from '@mantine/core';
 import { IconBold, IconItalic, IconSubscript, IconSuperscript, IconUnderline } from '@tabler/icons-react';
 import { useEffect, useState } from 'react';
-import { PhoenixObject, createReflection, updateReflection } from '../reflection/helpers';
+import Reflection from '../reflection';
 import Shadow from '../shadow';
 import CustomFont from './CustomFont';
 import { Font } from './types';
@@ -17,7 +17,6 @@ const TextPanel = ({ canvas, currentSelectedElement, artboardRef }: PanelProps) 
 	const [value, setValue] = useState('');
 	const [fontWeight, setFontWeight] = useState('regular');
 	const currentFont = fontList.find(font => font.family === value) as Font;
-	const [reflection, setReflection] = useState(false);
 
 	useEffect(() => {
 		fetch('/fonts.json')
@@ -27,55 +26,6 @@ const TextPanel = ({ canvas, currentSelectedElement, artboardRef }: PanelProps) 
 			})
 			.catch(error => console.error(error));
 	}, []);
-
-	useEffect(() => {
-		const element = currentSelectedElement?.[0] as PhoenixObject;
-		const reflection = (currentSelectedElement?.[0] as PhoenixObject).reflection as fabric.Textbox;
-		if (reflection && element.type === 'textbox') {
-			element.on('resizing', () => {
-				reflection.width = element.width;
-				reflection.height = element.height;
-				updateReflection(element, canvas);
-			});
-
-			element.on('changed', () => {
-				reflection.text = (element as fabric.Textbox).text;
-				canvas.requestRenderAll();
-				updateReflection(element, canvas);
-			});
-		}
-	}, [canvas, currentSelectedElement]);
-
-	useEffect(() => {
-		if ((currentSelectedElement?.[0] as PhoenixObject).reflection) {
-			setReflection(true);
-		}
-	}, [currentSelectedElement]);
-
-	useEffect(() => {
-		if (reflection) {
-			currentSelectedElement?.[0].on('moving', () => {
-				console.log('moving');
-				updateReflection(currentSelectedElement?.[0], canvas);
-			});
-
-			currentSelectedElement?.[0].on('scaling', () => {
-				console.log('scaling');
-				updateReflection(currentSelectedElement?.[0], canvas);
-			});
-
-			currentSelectedElement?.[0].on('rotating', event => {
-				console.log('rotating', event);
-				updateReflection(currentSelectedElement?.[0], canvas);
-			});
-		}
-
-		return () => {
-			currentSelectedElement?.[0].off('moving');
-			currentSelectedElement?.[0].off('scaling');
-			currentSelectedElement?.[0].off('rotating');
-		};
-	}, [currentSelectedElement, canvas, reflection]);
 
 	const toggleBold = () => {
 		const textElement = currentSelectedElement?.[0] as fabric.Text;
@@ -225,35 +175,7 @@ const TextPanel = ({ canvas, currentSelectedElement, artboardRef }: PanelProps) 
 				<Shadow canvas={canvas} currentSelectedElement={currentSelectedElement[0]} artboardRef={artboardRef} />
 			</Stack>
 			<Stack>
-				<Checkbox
-					label="Reflection"
-					checked={reflection}
-					onChange={e => {
-						setReflection(e.currentTarget.checked);
-						console.log('Checked?', e.currentTarget.checked);
-						if (e.currentTarget.checked) {
-							console.log('Creating reflection', currentSelectedElement?.[0]);
-							createReflection(currentSelectedElement?.[0] as PhoenixObject, canvas);
-							canvas.requestRenderAll();
-						} else {
-							console.log('Removing reflection', currentSelectedElement?.[0]);
-							// Remove reflection from canvas
-							const reflection = canvas
-								.getObjects()
-								.find(
-									object =>
-										object.data.type === 'reflection' &&
-										object.data.parent === currentSelectedElement?.[0].data.id,
-								);
-							if (reflection) {
-								canvas.remove(reflection);
-							}
-							// Remove reflection from object
-							(currentSelectedElement?.[0] as PhoenixObject).reflection = null;
-							canvas.requestRenderAll();
-						}
-					}}
-				/>
+				<Reflection canvas={canvas} currentSelectedElement={currentSelectedElement} />
 			</Stack>
 		</Stack>
 	);
